chore: Renaming api/index.js to api/app.js for better consistency in the use of index.js
