+++ src/shared/api/createUserCredentials.js
@@ -1,4 +1,4 @@
-import { app } from "./";
+import { app } from "./app";
 
 export const createUserCredentials = async (body) => {
   return app.service("user-credentials").create(body);
